リファクタ #82
